enterprise/admin-console/server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
boto3>=1.34.0
requests>=2.31.0
PyJWT[crypto]>=2.8.0
bcrypt>=4.0.0
